+++ samples/irs-demo/src/main/kotlin/net/corda/irs/Main.kt
@@ -7,8 +7,8 @@ import net.corda.core.utilities.DUMMY_BANK_A
 import net.corda.core.utilities.DUMMY_BANK_B
 import net.corda.core.utilities.DUMMY_NOTARY
 import net.corda.irs.api.NodeInterestRates
-import net.corda.node.driver.driver
 import net.corda.node.services.transactions.SimpleNotaryService
+import net.corda.testing.driver.driver
 
 /**
  * This file is exclusively for being able to run your nodes through an IDE (as opposed to running deployNodes)