+++ verifier/src/integration-test/kotlin/net/corda/verifier/VerifierDriver.kt
@@ -15,7 +15,6 @@ import net.corda.core.random63BitValue
 import net.corda.core.transactions.LedgerTransaction
 import net.corda.core.utilities.ProcessUtilities
 import net.corda.core.utilities.loggerFor
-import net.corda.node.driver.*
 import net.corda.node.services.config.configureDevKeyAndTrustStores
 import net.corda.nodeapi.ArtemisMessagingComponent.Companion.NODE_USER
 import net.corda.nodeapi.ArtemisTcpTransport
@@ -23,6 +22,7 @@ import net.corda.nodeapi.ConnectionDirection
 import net.corda.nodeapi.VerifierApi
 import net.corda.nodeapi.config.NodeSSLConfiguration
 import net.corda.nodeapi.config.SSLConfiguration
+import net.corda.testing.driver.*
 import org.apache.activemq.artemis.api.core.SimpleString
 import org.apache.activemq.artemis.api.core.client.ActiveMQClient
 import org.apache.activemq.artemis.api.core.client.ClientProducer